backend/content-service/requirements.txt
Flask==2.2.2
Werkzeug==2.2.3
flask-cors==5.0.1
pymongo==4.3.3
python-dotenv==0.19.2